Stop ignoring the dconf user.ini
Ignoring the user.ini sounds like it makes sense, but it actually makes management harder. Keeping it in both the source and destination allows me to unconditionally dump the current dconf state to dconf.ini in the destination and diff the destination with target state to see what changes should be added to the source state.